Use `bootstrapApplication`
