+++ client/src/app/openslides-main-module/components/openslides-main/openslides-main.component.ts
@@ -1,13 +1,7 @@
-import {
-    ApplicationRef,
-    ChangeDetectionStrategy,
-    Component,
-    DOCUMENT,
-    inject,
-    OnInit,
-    ViewContainerRef
-} from '@angular/core';
+import { ApplicationRef, Component, DOCUMENT, inject, OnInit } from '@angular/core';
 import { MatIconRegistry } from '@angular/material/icon';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { BrowserModule } from '@angular/platform-browser';
 import { DomSanitizer } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { allAvailableTranslations } from '@app/domain/definitions/languages';
@@ -16,9 +10,12 @@ import { StorageService } from '@app/gateways/storage.service';
 import { langToTimeLocale } from '@app/infrastructure/utils';
 import { overloadJsFunctions } from '@app/infrastructure/utils/overload-js-functions';
 import { Deferred } from '@app/infrastructure/utils/promises';
+import { OpenSlidesMainRoutingModule } from '@app/openslides-main-module/openslides-main-routing.module';
 import { BaseViewModel } from '@app/site/base/base-view-model';
+import { GlobalSpinnerModule } from '@app/site/modules/global-spinner';
 import { UpdateService } from '@app/site/modules/site-wrapper/services/update.service';
 import { CustomTranslationService } from '@app/site/modules/translations/custom-translation.service';
+import { WaitForActionDialogModule } from '@app/site/modules/wait-for-action-dialog';
 import { LifecycleService } from '@app/site/services/lifecycle.service';
 import { OpenSlidesService } from '@app/site/services/openslides.service';
 import { OpenSlidesStatusService } from '@app/site/services/openslides-status.service';
@@ -27,14 +24,22 @@ import { LangChangeEvent, TranslateService } from '@ngx-translate/core';
 import { DateFnsConfigurationService } from 'ngx-date-fns';
 import { first, firstValueFrom, tap } from 'rxjs';
 
+import { OpenSlidesOverlayContainerComponent } from '../openslides-overlay-container/openslides-overlay-container.component';
+
 const CURRENT_LANGUAGE_STORAGE_KEY = `currentLanguage`;
 
 @Component({
     selector: `os-root`,
     templateUrl: `./openslides-main.component.html`,
     styleUrls: [`./openslides-main.component.scss`],
-    changeDetection: ChangeDetectionStrategy.Eager,
-    standalone: false
+    imports: [
+        BrowserModule,
+        OpenSlidesMainRoutingModule,
+        OpenSlidesOverlayContainerComponent,
+        MatSnackBarModule,
+        GlobalSpinnerModule,
+        WaitForActionDialogModule
+    ]
 })
 export class OpenSlidesMainComponent implements OnInit {
     private onInitDone = new Deferred();
@@ -55,7 +60,9 @@ export class OpenSlidesMainComponent implements OnInit {
     private modelStore = inject(ViewModelStoreService);
     private ctService = inject(CustomTranslationService);
 
-    public constructor(_viewContainer: ViewContainerRef, _openslidesService: OpenSlidesService) {
+    public constructor() {
+        inject(OpenSlidesService);
+
         overloadJsFunctions();
         this.addDebugFunctions();
         this.waitForAppLoaded();
